test time delta tracking

+++ acmApp/src/acm_drv.h
@@ -148,7 +148,7 @@ struct Driver : public epicsThreadRunable
     sequences_t sequences;
 
     uint32_t lastTimebase;
-    typedef std::vector<std::pair<uint32_t, double> > tbhist_t;
+    typedef std::vector<std::pair<double, double> > tbhist_t;
     tbhist_t tbhist;
 
     // all other data members guarded by lock
